2.0.0

+++ scripts/remove_mongodb.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
 # MongoDB Uninstaller
-# Min. Requirement  : GNU/Linux Ubuntu 14.04
+# Min. Requirement  : GNU/Linux Ubuntu 16.04
 # Last Build        : 06/11/2019
-# Author            : ESLabs.ID (eslabs.id@gmail.com)
+# Author            : MasEDI.Net (me@masedi.net)
 # Since Version     : 1.0.0
 
 # Include helper functions.
@@ -20,7 +20,7 @@ requires_root
 function init_mongodb_removal() {
     # Stop MongoDB server process.
     if [[ $(pgrep -c mongod) -gt 0 ]]; then
-        run service mongod stop
+        run systemctl stop mongod
     fi
 
     if dpkg-query -l | awk '/mongodb/ { print $2 }' | grep -qwE "^mongodb"; then
@@ -28,9 +28,10 @@ function init_mongodb_removal() {
 
         # Remove MongoDB server.
         #shellcheck disable=SC2046
-        run apt-get -qq --purge remove -y $(dpkg-query -l | awk '/mongodb/ { print $2 }' | grep -wE "^mongodb")
-        run rm -f /etc/apt/sources.list.d/mongodb-org-*
-        run apt-get -qq autoremove -y
+        run apt remove --purge -qq -y $(dpkg-query -l | awk '/mongodb/ { print $2 }')
+        if "${FORCE_REMOVE}"; then
+            run rm -f /etc/apt/sources.list.d/mongodb-org-*
+        fi
 
         # Remove MongoDB config files.
         warning "!! This action is not reversible !!"
@@ -60,12 +61,12 @@ function init_mongodb_removal() {
 
     # Final test.
     if "${DRYRUN}"; then
-        warning "MongoDB server removed in dryrun mode."
+        info "MongoDB server removed in dryrun mode."
     else
         if [[ -z $(command -v mongod) ]]; then
-            status "MongoDB server removed succesfully."
+            success "MongoDB server removed succesfully."
         else
-            warning "Unable to remove MongoDB server."
+            info "Unable to remove MongoDB server."
         fi
     fi
 }
@@ -86,5 +87,5 @@ if [[ -n $(command -v mongod) ]]; then
         echo "Found MongoDB server, but not removed."
     fi
 else
-    warning "Oops, MongoDB installation not found."
+    info "Oops, MongoDB installation not found."
 fi
