+++ scripts/remove_nginx.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
-# Nginx uninstaller
-# Min. Requirement  : GNU/Linux Ubuntu 14.04
+# NGiNX uninstaller
+# Min. Requirement  : GNU/Linux Ubuntu 16.04
 # Last Build        : 31/07/2019
-# Author            : ESLabs.ID (eslabs.id@gmail.com)
+# Author            : MasEDI.Net (me@masedi.net)
 # Since Version     : 1.0.0
 
 # Include helper functions.
@@ -21,80 +21,75 @@ requires_root
 function init_nginx_removal() {
     # Stop nginx HTTP server process.
     if [[ $(pgrep -c nginx) -gt 0 ]]; then
-        run service nginx stop
+        #run service nginx stop
+        run systemctl stop nginx
+    fi
+
+    if [[ ${NGX_VERSION} == "mainline" || ${NGX_VERSION} == "latest" ]]; then
+        local NGINX_REPO="nginx-mainline"
+    else
+        local NGINX_REPO="nginx"
     fi
 
     # Remove nginx installation.
     if dpkg-query -l | awk '/nginx/ { print $2 }' | grep -qwE "^nginx-stable"; then
-        echo "Nginx-common package found. Removing..."
-        run apt-get -qq --purge remove -y nginx-stable
+        echo "Found nginx-stable package installation, removing..."
+
+        # shellcheck disable=SC2046
+        run apt remove --purge -qq -y $(dpkg-query -l | awk '/nginx/ { print $2 }' | grep -wE "^nginx")
         if "${FORCE_REMOVE}"; then
             run add-apt-repository -y --remove ppa:nginx/stable
         fi
     elif dpkg-query -l | awk '/nginx/ { print $2 }' | grep -qwE "^nginx-custom"; then
-        echo "Nginx-custom package found. Removing..."
-        run apt-get -qq --purge remove -y nginx-custom
+        echo "Found nginx-custom package installation, removing..."
+
+        # shellcheck disable=SC2046
+        run apt remove --purge -qq -y $(dpkg-query -l | awk '/nginx/ { print $2 }' | grep -wE "^nginx")
         if "${FORCE_REMOVE}"; then
             run add-apt-repository -y --remove ppa:rtcamp/nginx
         fi
-    elif dpkg-query -l | awk '/nginx/ { print $2 }' | grep -qwE "^nginx-extras"; then
-        echo "Nginx-stable package found. Removing..."
-        run apt-get -qq --purge remove -y nginx-extras
+    elif dpkg-query -l | awk '/nginx/ { print $2 }' | grep -qwE "^nginx"; then
+        echo "Found nginx package installation, removing..."
+
+        # shellcheck disable=SC2046
+        run apt remove --purge -qq -y $(dpkg-query -l | awk '/nginx/ { print $2 }' | grep -wE "^nginx") $(dpkg-query -l | awk '/libnginx/ { print $2 }' | grep -wE "^libnginx")
         if "${FORCE_REMOVE}"; then
-            run add-apt-repository -y --remove ppa:ondrej/nginx
+            run add-apt-repository -y --remove "ppa:ondrej/${NGINX_REPO}"
         fi
     else
-        echo "Nginx package not found. Possibly installed from source."
+        echo "No installed nginx package found, possibly installed from source."
 
         NGINX_BIN=$(command -v nginx)
 
-        echo "Nginx binary executable: ${NGINX_BIN}"
-
         if [[ -n "${NGINX_BIN}" ]]; then
+            echo "NGiNX binary executable: ${NGINX_BIN}"
+
             # Disable systemctl.
-            if [ -f /etc/systemd/system/multi-user.target.wants/nginx.service ]; then
-                echo "Disable Nginx service..."
-                run systemctl disable nginx.service
-            fi
+            echo "Disable NGiNX service..."
+            [ -f /etc/systemd/system/multi-user.target.wants/nginx.service ] && run systemctl disable nginx
+            [ -f /etc/systemd/system/multi-user.target.wants/nginx.service ] && \
+            run unlink /etc/systemd/system/multi-user.target.wants/nginx.service
+            [ -f /lib/systemd/system/nginx.service ] && run rm -f /lib/systemd/system/nginx.service
 
-            if [ -f /etc/systemd/system/multi-user.target.wants/nginx.service ]; then
-                run unlink /etc/systemd/system/multi-user.target.wants/nginx.service
-            fi
-
-            if [ -f /lib/systemd/system/nginx.service ]; then
-                run rm -f /lib/systemd/system/nginx.service
-            fi
+            # Remove Nginx files.
+            [ -f /etc/init.d/nginx ] && run rm -f /etc/init.d/nginx
+            [ -d /usr/lib/nginx ] && run rm -fr /usr/lib/nginx
+            [ -d /etc/nginx/modules-enabled ] && run rm -fr /etc/nginx/modules-enabled
+            [ -d /etc/nginx/modules-available ] && run rm -fr /etc/nginx/modules-available
 
             # Remove binary executable file.
-            if [ -f "${NGINX_BIN}" ]; then
-                echo "Remove Nginx binary executable file..."
+            if [ -x "${NGINX_BIN}" ]; then
+                echo "Remove NGiNX binary executable file..."
                 run rm -f "${NGINX_BIN}"
             fi
 
-            # Remove init file.
-            if [ -f /etc/init.d/nginx ]; then
-                run rm -f /etc/init.d/nginx
-            fi
-
-            if [ -d /usr/lib/nginx/ ]; then
-                run rm -fr /usr/lib/nginx/
-            fi
-
-            if [ -d /etc/nginx/modules-enabled ]; then
-                run rm -fr /etc/nginx/modules-enabled
-            fi
-
-            if [ -d /etc/nginx/modules-available ]; then
-                run rm -fr /etc/nginx/modules-available
-            fi
-
-            # Delete default account from PageSpeed admin.
+            # Delete default account credential from server .htpasswd.
             if [ -f /srv/.htpasswd ]; then
                 local USERNAME=${LEMPER_USERNAME:-"lemper"}
-                #run echo "" > /srv/.htpasswd
-                #run sed -i "/^lemper:/d" /srv/.htpasswd
                 run sed -i "/^${USERNAME}:/d" /srv/.htpasswd
             fi
+        else
+            error "Sorry, we couldn't find any NGiNX binary executable file."
         fi
     fi
 
@@ -104,40 +99,38 @@ function init_nginx_removal() {
         REMOVE_NGXCONFIG="y"
     else
         while [[ "${REMOVE_NGXCONFIG}" != "y" && "${REMOVE_NGXCONFIG}" != "n" ]]; do
-            read -rp "Remove all Nginx configuration files? [y/n]: " -e REMOVE_NGXCONFIG
+            read -rp "Remove all NGiNX configuration files? [y/n]: " -e REMOVE_NGXCONFIG
         done
     fi
 
     if [[ "${REMOVE_NGXCONFIG}" == Y* || "${REMOVE_NGXCONFIG}" == y* || "${FORCE_REMOVE}" == true ]]; then
         run rm -fr /etc/nginx
-        # Remove nginx-cache.
         run rm -fr /var/cache/nginx
-        # Remove nginx html.
         run rm -fr /usr/share/nginx
 
-        echo "All your Nginx configs files deleted permanently."
+        echo "All your NGiNX configuration files deleted permanently."
     fi
     
     # Final test.
     if "${DRYRUN}"; then
-        warning "Nginx HTTP server removed in dryrun mode."
+        info "NGiNX HTTP server removed in dryrun mode."
     else
         if [[ -z $(command -v nginx) ]]; then
-            status "Nginx HTTP server removed succesfully."
+            success "NGiNX HTTP server removed succesfully."
         else
-            warning "Unable to remove Nginx HTTP server."
+            info "Unable to remove NGiNX HTTP server."
         fi
     fi
 }
 
-echo "Uninstalling Nginx HTTP server..."
-if [[ -n $(command -v nginx) || -x /usr/sbin/nginx ]]; then
+echo "Uninstalling NGiNX HTTP server..."
 
+if [[ -n $(command -v nginx) || -x /usr/sbin/nginx ]]; then
     if "${AUTO_REMOVE}"; then
         REMOVE_NGINX="y"
     else
         while [[ "${REMOVE_NGINX}" != "y" && "${REMOVE_NGINX}" != "n" ]]; do
-            read -rp "Are you sure to remove Nginx HTTP server? [y/n]: " -e REMOVE_NGINX
+            read -rp "Are you sure to remove NGiNX HTTP server? [y/n]: " -e REMOVE_NGINX
         done
     fi
 
@@ -147,5 +140,5 @@ if [[ -n $(command -v nginx) || -x /usr/sbin/nginx ]]; then
         echo "Found NGiNX HTTP server, but not removed."
     fi
 else
-    warning "Oops, NGiNX installation not found."
+    info "Oops, NGiNX installation not found."
 fi
